registration with auth and working
